refactor(prepare-analysis): decompose PromptGenerator into façade + 4 mixins
Extract 20 methods from prepare_analysis.py (1500 LOC) into 4 responsibility-based
mixins under workflows/prompt/: MetricHelpersMixin, ContextLoadingMixin,
DataFormattingMixin, PromptAssemblyMixin. Façade retains __init__ + 3 module-level
functions. 6th god script refactored in Phase 3 (111 tests pass, 0 regressions).
