+++ magma_cycling/prepare_analysis.py
@@ -54,7 +54,6 @@ Metadata:
     Version: v2
 """
 import argparse
-import json
 import subprocess
 import sys
 from datetime import datetime, timedelta
@@ -65,9 +64,18 @@ import requests
 from magma_cycling.api.intervals_client import IntervalsClient
 from magma_cycling.config import create_intervals_client, get_ai_config
 from magma_cycling.workflow_state import WorkflowState
+from magma_cycling.workflows.prompt.context_loading import ContextLoadingMixin
+from magma_cycling.workflows.prompt.data_formatting import DataFormattingMixin
+from magma_cycling.workflows.prompt.metric_helpers import MetricHelpersMixin
+from magma_cycling.workflows.prompt.prompt_assembly import PromptAssemblyMixin
 
 
-class PromptGenerator:
+class PromptGenerator(
+    ContextLoadingMixin,
+    DataFormattingMixin,
+    MetricHelpersMixin,
+    PromptAssemblyMixin,
+):
     """Générateur de prompt pour analyse IA."""
 
     def __init__(self, project_root=None):
@@ -101,833 +109,6 @@ class PromptGenerator:
 
         self.feedback_dir = Path(".athlete_feedback")
         self.feedback_file = self.feedback_dir / "last_feedback.json"
-
-    def load_athlete_context(self):
-        """Load le contexte athlète depuis project_prompt_v2_1_revised.md."""
-        prompt_file = self.references_dir / "project_prompt_v2_1_revised.md"
-
-        if prompt_file.exists():
-            with open(prompt_file, encoding="utf-8") as f:
-                return f.read()
-        return None
-
-    def load_recent_workouts(self, limit=5):
-        """Load les N dernières séances depuis workouts-history.md."""
-        history_file = self.logs_dir / "workouts-history.md"
-
-        if not history_file.exists():
-            return None
-
-        with open(history_file, encoding="utf-8") as f:
-            content = f.read()
-
-        # Extraire les dernières entrées (simplifié)
-        # On cherche les sections ### et prend les N premières après "## Historique"
-        sections = content.split("###")
-        recent = []
-        count = 0
-        for section in sections:
-            if count >= limit:
-                break
-            if section.strip() and "Date :" in section:
-                recent.append("###" + section)
-                count += 1
-
-        return "\n".join(recent) if recent else None
-
-    def load_cycling_concepts(self):
-        """Load les concepts d'entraînement cyclisme."""
-        concepts_file = self.references_dir / "cycling_training_concepts.md"
-
-        if concepts_file.exists():
-            with open(concepts_file, encoding="utf-8") as f:
-                return f.read()
-        return None
-
-    def load_periodization_context(self, wellness_data: dict | None = None) -> dict | None:
-        """
-        Load periodization context (macro/micro-cycle position).
-
-        Provides strategic context about training phase, CTL progression,
-        and PID controller state for more coherent AI analysis.
-
-        Args:
-            wellness_data: Wellness data containing CTL metrics
-
-        Returns:
-            Dict with periodization context or None if unavailable
-
-        Examples:
-            >>> context = generator.load_periodization_context(wellness_data)
-            >>> print(context['phase'])  # "RECONSTRUCTION_BASE"
-            >>> print(context['ctl_target'])  # 73.0
-        """
-        from magma_cycling.config.athlete_profile import AthleteProfile
-        from magma_cycling.planning.peaks_phases import determine_training_phase
-        from magma_cycling.workflows.pid_peaks_integration import (
-            ControlMode,
-            compute_integrated_correction,
-        )
-
-        try:
-            # Load athlete profile
-            profile = AthleteProfile.from_env()
-
-            # Extract CTL from wellness data
-            if not wellness_data:
-                return None
-
-            from magma_cycling.utils.metrics import extract_wellness_metrics
-
-            metrics = extract_wellness_metrics(wellness_data)
-            ctl_current = metrics["ctl"]
-
-            if ctl_current == 0:
-                return None
-
-            # Determine training phase
-            phase_rec = determine_training_phase(
-                ctl_current=ctl_current,
-                ftp_current=profile.ftp,
-                ftp_target=profile.ftp_target,
-                athlete_age=profile.age,
-            )
-
-            # Compute PID integration status
-            integrated = compute_integrated_correction(
-                ctl_current=ctl_current,
-                ftp_current=profile.ftp,
-                ftp_target=profile.ftp_target,
-                athlete_age=profile.age,
-            )
-
-            # Determine PID status
-            if integrated.override_active:
-                pid_status = "Override Peaks (CTL < 50 - reconstruction urgente)"
-            elif integrated.mode == ControlMode.PID_CONSTRAINED:
-                pid_status = "Actif avec contraintes Peaks"
-            else:
-                pid_status = "Actif (autonome)"
-
-            # Calculate weeks to rebuild (approximate)
-            weeks_to_target = phase_rec.weeks_to_rebuild if phase_rec.weeks_to_rebuild > 0 else 0
-
-            # Format phase label
-            phase_labels = {
-                "reconstruction_base": "RECONSTRUCTION BASE",
-                "consolidation": "CONSOLIDATION",
-                "development_ftp": "DÉVELOPPEMENT FTP",
-            }
-            phase_label = phase_labels.get(phase_rec.phase.value, phase_rec.phase.value.upper())
-
-            return {
-                "phase": phase_label,
-                "phase_raw": phase_rec.phase.value,
-                "ctl_current": ctl_current,
-                "ctl_target": phase_rec.ctl_target,
-                "ctl_deficit": phase_rec.ctl_deficit,
-                "ftp_current": profile.ftp,
-                "ftp_target": profile.ftp_target,
-                "weeks_to_target": weeks_to_target,
-                "pid_status": pid_status,
-                "pid_override_active": integrated.override_active,
-                "weekly_tss_load": phase_rec.weekly_tss_load,
-                "weekly_tss_recovery": phase_rec.weekly_tss_recovery,
-                "recovery_week_frequency": phase_rec.recovery_week_frequency,
-                "rationale": phase_rec.rationale,
-            }
-
-        except Exception as e:
-            # Fail gracefully if periodization context unavailable
-            print(f"⚠️  Impossible de charger le contexte de périodisation : {e}")
-            return None
-
-    def load_athlete_feedback(self):
-        """Load le feedback athlète s'il existe."""
-        if not self.feedback_file.exists():
-            return None
-
-        try:
-            with open(self.feedback_file, encoding="utf-8") as f:
-                return json.load(f)
-        except Exception as e:
-            print(f"⚠️  Erreur lecture feedback : {e}")
-            return None
-
-    def format_athlete_feedback(self, feedback):
-        """Format le feedback pour inclusion dans le prompt."""
-        if not feedback:
-            return None
-
-        parts = []
-
-        if feedback.get("rpe"):
-            parts.append(f"**RPE** : {feedback['rpe']}/10")
-
-        if feedback.get("ressenti_general"):
-            parts.append(f"**Ressenti** : {feedback['ressenti_general']}")
-
-        if feedback.get("difficultes"):
-            parts.append(f"**Difficultés** :\n{feedback['difficultes']}")
-
-        if feedback.get("points_positifs"):
-            parts.append(f"**Points positifs** :\n{feedback['points_positifs']}")
-
-        if feedback.get("contexte"):
-            parts.append(f"**Contexte** : {feedback['contexte']}")
-
-        if feedback.get("sensations_physiques"):
-            sensations = ", ".join(feedback["sensations_physiques"])
-            parts.append(f"**Sensations physiques** : {sensations}")
-
-        if feedback.get("notes_libres"):
-            parts.append(f"**Notes libres** :\n{feedback['notes_libres']}")
-
-        return "\n\n".join(parts) if parts else None
-
-    def format_activity_data(self, activity):
-        """Format les données d'activité pour le prompt."""
-        date = datetime.fromisoformat(activity["start_date_local"].replace("Z", "+00:00"))
-
-        # Vérifier si l'activité vient de Strava
-        is_strava = activity.get("source") == "STRAVA"
-
-        data = {
-            "id": activity.get("id", "Non spécifié"),
-            "name": activity.get("name", "Séance"),
-            "type": activity.get("type", "Cyclisme"),
-            "date": date.strftime("%d/%m/%Y"),
-            "date_iso": date.strftime("%Y-%m-%d"),
-            "duration_min": activity.get("moving_time", 0) // 60,
-            "tss": activity.get("icu_training_load", 0),
-            "intensity": activity.get("icu_intensity", 0) / 100.0,
-            "avg_power": activity.get("icu_average_watts", 0),
-            "np": activity.get("icu_weighted_avg_watts", 0),
-            "avg_cadence": activity.get("average_cadence", 0),
-            "avg_hr": activity.get("average_heartrate", 0),
-            "max_hr": activity.get("max_heartrate", 0),
-            "decoupling": activity.get("decoupling", None),
-            "avg_temp": activity.get("average_temp"),  # Temperature in °C
-            "min_temp": activity.get("min_temp"),
-            "max_temp": activity.get("max_temp"),
-            "has_weather": activity.get("has_weather", False),
-            "description": activity.get("description", ""),
-            "tags": activity.get("tags", []),
-            "feel": activity.get("feel"),  # 1-5 scale Intervals.icu (1=Excellent, 5=Poor)
-            "is_strava": is_strava,
-            "source": activity.get("source", "Unknown"),
-        }
-
-        return data
-
-    def _format_feel_value(self, feel_value):
-        """Format the 'feel' value (Intervals.icu 1-5 scale) into readable text.
-
-        Args:
-            feel_value: Numeric value 1-5 or None (1=Excellent, 5=Poor)
-
-        Returns:
-            Formatted string with emoji
-        """
-        if feel_value is None:
-            return "_Non renseigné_"
-
-        feel_map = {
-            1: "😊 Excellent (1/5)",
-            2: "🙂 Bien (2/5)",
-            3: "😐 Moyen (3/5)",
-            4: "😕 Passable (4/5)",
-            5: "😣 Mauvais (5/5)",
-        }
-
-        return feel_map.get(feel_value, f"_Valeur inconnue: {feel_value}_")
-
-    def _format_athlete_notes(self, activity_description, wellness_comments):
-        """Format athlete notes with wellness comments as fallback.
-
-        Args:
-            activity_description: Notes from activity description field
-            wellness_comments: Notes from wellness comments field
-
-        Returns:
-            Formatted string with notes and source indication
-        """
-        # Priority 1: Activity description
-        if activity_description and activity_description.strip():
-            return activity_description.strip()
-
-        # Priority 2: Wellness comments (fallback)
-        if wellness_comments and wellness_comments.strip():
-            return f"{wellness_comments.strip()}\n\n_Note: Feedback saisi dans wellness (pas de description activité)_"
-
-        # No notes at all
-        return "_Aucune note saisie_"
-
-    def _format_temperature_data(self, avg_temp, min_temp, max_temp, has_weather):
-        """Format temperature data for activity analysis.
-
-        Args:
-            avg_temp: Average temperature in °C
-            min_temp: Minimum temperature in °C
-            max_temp: Maximum temperature in °C
-            has_weather: Whether weather data is available
-
-        Returns:
-            Formatted string with temperature information
-        """
-        if not has_weather or avg_temp is None:
-            return "_Données météo non disponibles_"
-
-        # Format average temperature
-        avg_str = f"{avg_temp:.1f}°C"
-
-        # Add min/max if available
-        if min_temp is not None and max_temp is not None:
-            range_str = f" (min {min_temp}°C, max {max_temp}°C)"
-        else:
-            range_str = ""
-
-        # Add contextual emoji based on temperature
-        if avg_temp < 5:
-            emoji = "🥶"
-            context = "très froid"
-        elif avg_temp < 10:
-            emoji = "❄️"
-            context = "froid"
-        elif avg_temp < 15:
-            emoji = "🌡️"
-            context = "frais"
-        elif avg_temp < 20:
-            emoji = "☀️"
-            context = "tempéré"
-        elif avg_temp < 25:
-            emoji = "🌤️"
-            context = "agréable"
-        elif avg_temp < 30:
-            emoji = "☀️"
-            context = "chaud"
-        else:
-            emoji = "🔥"
-            context = "très chaud"
-
-        return f"{emoji} {avg_str}{range_str} ({context})"
-
-    def format_wellness_data(self, wellness):
-        """Format les données wellness."""
-        if not wellness:
-            return {
-                "ctl": 0,
-                "atl": 0,
-                "tsb": 0,
-                "weight": 0,
-                "sleep_seconds": 0,
-                "comments": "",
-            }
-
-        from magma_cycling.utils.metrics import extract_wellness_metrics
-
-        metrics = extract_wellness_metrics(wellness)
-
-        return {
-            "ctl": metrics["ctl"],
-            "atl": metrics["atl"],
-            "tsb": metrics["tsb"],
-            "weight": wellness.get("weight", 0),
-            "sleep_seconds": wellness.get("sleepSecs", 0),
-            "sleep_quality": wellness.get("sleepQuality", 0),
-            "comments": wellness.get("comments", ""),
-        }
-
-    def format_planned_workout(self, planned_event):
-        """Format le workout planifié pour le prompt.
-
-        Args:
-            planned_event: L'événement contenant le workout planifié
-
-        Returns:
-            Dict avec les informations formatées ou None.
-        """
-        if not planned_event or not planned_event.get("workout_doc"):
-            return None
-
-        workout_doc = planned_event["workout_doc"]
-
-        # Extraire les données principales
-        formatted = {
-            "name": planned_event.get("name", "Workout planifié"),
-            "description": planned_event.get("description", ""),
-            "duration_min": workout_doc.get("duration", 0) // 60,
-            "tss_planned": planned_event.get("icu_training_load", 0),
-            "avg_watts_planned": workout_doc.get("average_watts", 0),
-            "np_planned": workout_doc.get("normalized_power", 0),
-            "intensity_planned": (
-                planned_event.get("icu_intensity", 0) / 100.0
-                if planned_event.get("icu_intensity")
-                else 0
-            ),
-            "joules": workout_doc.get("joules", 0),
-        }
-
-        # Formater la structure des intervalles
-        steps = workout_doc.get("steps", [])
-        intervals = []
-
-        for _i, step in enumerate(steps):
-            if "reps" in step:
-                # C'est un bloc d'intervalles répétés
-                reps = step["reps"]
-                sub_steps = step.get("steps", [])
-                interval_desc = []
-                for sub in sub_steps:
-                    power_info = self._format_power(sub.get("power", {}))
-                    cadence = sub.get("cadence", {}).get("value", 0)
-                    duration_min = sub.get("duration", 0) / 60
-                    interval_desc.append(f"{duration_min:.0f}min @ {power_info} / {cadence}rpm")
-                intervals.append(f"{reps}x ({' → '.join(interval_desc)})")
-            else:
-                # Step simple
-                power_info = self._format_power(step.get("power", {}))
-                cadence = step.get("cadence", {}).get("value", 0)
-                duration_min = step.get("duration", 0) / 60
-                step_type = ""
-                if step.get("warmup"):
-                    step_type = "[Warmup] "
-                elif step.get("cooldown"):
-                    step_type = "[Cooldown] "
-                intervals.append(f"{step_type}{duration_min:.0f}min @ {power_info} / {cadence}rpm")
-
-        formatted["intervals"] = intervals
-
-        # Répartition des zones (time in zones)
-        zone_times = workout_doc.get("zoneTimes", [])
-        zones_str = []
-        for zone in zone_times:
-            if zone.get("secs", 0) > 0 and not zone.get("gap"):  # Ignorer Sweet Spot (gap=true)
-                zone_name = zone.get("name", zone.get("id", "Z?"))
-                zone_min = zone["secs"] / 60
-                zones_str.append(f"{zone_name}: {zone_min:.0f}min")
-
-        formatted["zone_distribution"] = ", ".join(zones_str) if zones_str else "N/A"
-
-        return formatted
-
-    def _format_power(self, power_dict):
-        """Format une valeur de puissance (gère %, watts absolus, rampes)."""
-        if not power_dict:
-            return "N/A"
-
-        units = power_dict.get("units", "")
-
-        if units == "%ftp":
-            if "value" in power_dict:
-                return f"{power_dict['value']}%FTP"
-            elif "start" in power_dict and "end" in power_dict:
-                return f"{power_dict['start']}-{power_dict['end']}%FTP"
-        elif units == "w":
-            return f"{power_dict.get('value', 0)}W"
-        elif units == "power_zone":
-            return f"Z{power_dict.get('value', '?')}"
-
-        return "N/A"
-
-    def safe_format_metric(self, value, format_spec=".0f", suffix="", default="N/A"):
-        """
-        Formatage sécurisé d'une métrique pouvant être None.
-
-        Args:
-            value: Valeur à formater (peut être None)
-            format_spec: Spécification de format (ex: '.0f', '.2f')
-            suffix: Suffixe à ajouter (ex: 'W', 'bpm')
-            default: Valeur par défaut si None
-
-        Returns:
-            str: Valeur formatée ou default.
-        """
-        if value is None:
-            return default
-        try:
-            return f"{value:{format_spec}}{suffix}"
-        except (ValueError, TypeError):
-            return default
-
-    def get_power_value(self, activity, metric_type="avg"):
-        """
-        Extract robuste de la puissance avec fallback multi-champs.
-
-        Args:
-            activity: Dictionnaire activité depuis API
-            metric_type: 'avg' (moyenne) ou 'np' (normalisée)
-
-        Returns:
-            float ou None: Valeur de puissance trouvée.
-        """
-        field_mappings = {
-            "avg": ["avg_power", "power", "average_power", "watts", "avgWatts"],
-            "np": ["np", "normalized_power", "norm_power", "normalizedPower"],
-        }
-
-        possible_fields = field_mappings.get(metric_type, [])
-
-        for field in possible_fields:
-            value = activity.get(field)
-            if value is not None and value > 0:
-                return value
-
-        return None
-
-    def get_cadence_value(self, activity, metric_type="avg"):
-        """Extract robuste de la cadence avec fallback."""
-        field_mappings = {
-            "avg": ["avg_cadence", "cadence", "avgCadence", "rpm"],
-            "max": ["max_cadence", "maxCadence", "maximum_cadence"],
-        }
-
-        possible_fields = field_mappings.get(metric_type, [])
-
-        for field in possible_fields:
-            value = activity.get(field)
-            if value is not None and value > 0:
-                return value
-
-        return None
-
-    def get_hr_value(self, activity, metric_type="avg"):
-        """Extract robuste de la fréquence cardiaque avec fallback."""
-        field_mappings = {
-            "avg": ["avg_hr", "hr", "avgHr", "heart_rate", "average_hr"],
-            "max": ["max_hr", "maxHr", "max_heart_rate", "maximum_hr"],
-        }
-
-        possible_fields = field_mappings.get(metric_type, [])
-
-        for field in possible_fields:
-            value = activity.get(field)
-            if value is not None and value > 0:
-                return value
-
-        return None
-
-    def generate_prompt(
-        self,
-        activity_data,
-        wellness_pre,
-        wellness_post,
-        athlete_context,
-        recent_workouts,
-        athlete_feedback=None,
-        planned_workout=None,
-        cycling_concepts=None,
-        periodization_context=None,
-        session_prescription=None,
-    ):
-        """Generate le prompt complet pour analyse IA."""
-        # Formater les données
-
-        act = activity_data
-        w_pre = self.format_wellness_data(wellness_pre)
-        w_post = self.format_wellness_data(wellness_post)
-        planned = self.format_planned_workout(planned_workout) if planned_workout else None
-
-        decoupling_str = f"{act['decoupling']:.1f}%" if act["decoupling"] else "N/A"
-
-        # Format temperature data
-        temperature_str = self._format_temperature_data(
-            act.get("avg_temp"),
-            act.get("min_temp"),
-            act.get("max_temp"),
-            act.get("has_weather", False),
-        )
-
-        # Extraction robuste des métriques avec fallback
-        avg_power = self.get_power_value(act, "avg")
-        normalized_power = self.get_power_value(act, "np")
-        avg_cadence = self.get_cadence_value(act, "avg")
-        avg_hr = self.get_hr_value(act, "avg")
-        max_hr = self.get_hr_value(act, "max")
-
-        # Warning si données puissance manquantes
-        power_warning = ""
-        if avg_power is None and normalized_power is None:
-            power_warning = "\n⚠️  ATTENTION : Aucune donnée de puissance disponible (fichier .fit incomplet ou séance sans capteur)\n   → Analyse basée sur FC, cadence et RPE uniquement\n\n"
-        sleep_hours = (
-            w_pre["sleep_seconds"] / 3600
-            if w_pre["sleep_seconds"] and w_pre["sleep_seconds"] > 0
-            else 0
-        )
-        weight_kg = w_pre["weight"] if w_pre["weight"] and w_pre["weight"] > 0 else 0
-
-        # Avertissement sommeil manquant
-        sleep_warning = ""
-        if sleep_hours == 0:
-            sleep_warning = "\n⚠️  ATTENTION : Données de sommeil non disponibles dans Intervals.icu au moment de l'analyse\n   → Saisir les données wellness (Sleep) dans Intervals.icu avant génération du prompt\n   → Ou fournir l'information dans le feedback athlète\n\n"
-
-        # Avertissement Strava si nécessaire
-        strava_warning = ""
-        if act["is_strava"]:
-            strava_warning = f"""
-⚠️  **ATTENTION : Activité Strava**
-
-Source : {act['source']}
-Les données API peuvent être limitées par les restrictions Strava.
-Certaines métriques (puissance, découplage) peuvent être manquantes ou incomplètes.
-→ Vérifier les données sur l'interface web Intervals.icu si nécessaire.
-
-"""
-        prompt = f"""# Analyse d'Entraînement Cyclisme.
-
-## Contexte Athlète
-
-{athlete_context if athlete_context else "[Contexte non disponible - utiliser informations par défaut]"}
-
----
-
-## 📚 Référence Cyclisme
-
-{cycling_concepts if cycling_concepts else "[Concepts cyclisme non disponibles]"}
-
----
-
-## Séance à Analyser
-
-{strava_warning}{power_warning}{sleep_warning}### Informations Générales
-- **Nom** : {act['name']}
-- **ID** : {act['id']}
-- **Type** : {act['type']}
-- **Date** : {act['date']}
-- **Source** : {act['source']}
-
-### Métriques Pré-séance
-- CTL : {w_pre['ctl']:.0f}
-- ATL : {w_pre['atl']:.0f}
-- TSB : {w_pre['tsb']:+.0f}
-- Poids : {weight_kg:.1f}kg
-- Sommeil : {sleep_hours:.1f}h
-
-### Exécution
-- Durée réalisée : {act['duration_min']}min
-- IF : {act['intensity']:.2f}
-- TSS : {self.safe_format_metric(act.get('tss'), '.0f', '')}
-- Puissance moyenne : {self.safe_format_metric(avg_power, '.0f', 'W')}
-- Puissance normalisée : {self.safe_format_metric(normalized_power, '.0f', 'W')}
-- Cadence moyenne : {self.safe_format_metric(avg_cadence, '.0f', 'rpm')}
-- FC moyenne : {self.safe_format_metric(avg_hr, '.0f', 'bpm')}
-- FC max : {self.safe_format_metric(max_hr, '.0f', 'bpm')}
-- Découplage cardiovasculaire : {decoupling_str}
-- Température : {temperature_str}
-
-### Métriques Post-séance
-- CTL : {w_post['ctl']:.0f}
-- ATL : {w_post['atl']:.0f}
-- TSB : {w_post['tsb']:+.0f}
-
-### Feedback Athlète (saisi dans Intervals.icu)
-
-**Ressenti général** : {self._format_feel_value(act.get('feel'))}
-
-**Notes athlète** :
-{self._format_athlete_notes(act.get('description', ''), w_pre.get('comments', ''))}
-
-**Tags** : {', '.join(act['tags']) if act['tags'] else '_Aucun tag_'}
-
----
-"""
-        # Ajouter prescription coach si disponible
-        if session_prescription:
-            prompt += f"""
-## 🎯 Prescription Coach (Objectifs de la Séance)
-
-{session_prescription}
-
-**Consigne d'analyse** : Évaluer si l'exécution répond aux objectifs prescrits
-ci-dessus. Identifier les écarts entre intention et réalisation.
-
----
-"""
-        # Ajouter section workout planifié si disponible
-
-        if planned:
-            prompt += f"""
-## 📋 Workout Planifié vs Réalisé
-
-### Objectifs Planifiés
-- **Nom** : {planned['name']}
-- **Durée prévue** : {planned['duration_min']}min
-- **TSS prévu** : {planned['tss_planned']:.0f}
-- **IF prévue** : {planned['intensity_planned']:.2f}
-- **Puissance moy. prévue** : {planned['avg_watts_planned']:.0f}W
-- **NP prévue** : {planned['np_planned']:.0f}W
-
-### Structure Planifiée
-"""
-            for interval in planned["intervals"]:
-                prompt += f"- {interval}\n"
-
-            prompt += f"""
-### Répartition Zones Planifiée
-
-{planned['zone_distribution']}
-
-### Description Workout
-{planned['description'][:500] if planned['description'] else '_Aucune description_'}{'...' if len(planned['description']) > 500 else ''}
-
-### Comparaison Planifié vs Réalisé
-- Durée : {planned['duration_min']}min prévu → {act['duration_min']}min réalisé ({act['duration_min'] - planned['duration_min']:+}min)
-- TSS : {planned['tss_planned']:.0f} prévu → {act['tss']:.0f} réalisé ({act['tss'] - planned['tss_planned']:+.0f})
-- IF : {planned['intensity_planned']:.2f} prévue → {act['intensity']:.2f} réalisée ({act['intensity'] - planned['intensity_planned']:+.2f})
-- Puissance moy. : {planned['avg_watts_planned']:.0f}W prévue → {act['avg_power']:.0f}W réalisée ({act['avg_power'] - planned['avg_watts_planned']:+.0f}W)
-- NP : {planned['np_planned']:.0f}W prévue → {act['np']:.0f}W réalisée ({act['np'] - planned['np_planned']:+.0f}W)
-
-**Consigne d'analyse** : Évaluer l'adhérence au plan et identifier les écarts significatifs (>10% en durée/TSS, >5% en IF).
-
----.
-"""
-        # Ajouter contexte de périodisation si disponible
-        if periodization_context:
-            pc = periodization_context
-            prompt += f"""
-
-## 📊 Contexte Périodisation (Macro/Micro-Cycle)
-
-### Phase Actuelle : {pc['phase']}
-
-**Objectifs Cycle** :
-- CTL actuel : {pc['ctl_current']:.1f}
-- CTL cible : {pc['ctl_target']:.0f}
-- Déficit CTL : {pc['ctl_deficit']:.1f} points
-- FTP actuel : {pc['ftp_current']}W
-- FTP cible : {pc['ftp_target']}W
-
-**Progression** :
-- Durée reconstruction estimée : {pc['weeks_to_target']} semaines
-- TSS semaines charge : {pc['weekly_tss_load']} TSS
-- TSS semaines récupération : {pc['weekly_tss_recovery']} TSS
-- Fréquence récupération : Tous les {pc['recovery_week_frequency']} semaines
-
-**État PID Controller** : {pc['pid_status']}
-
-**Rationale Phase** :
-{pc['rationale']}
-
-**➡️ Impact sur l'Analyse** : Les recommandations doivent tenir compte de la phase actuelle. En RECONSTRUCTION BASE, prioriser volume aérobie (Tempo/Sweet-Spot). En CONSOLIDATION, introduire progressivement FTP/VO2. En DÉVELOPPEMENT FTP, focus intensité.
-
----
-"""
-        prompt += f"""
-
-## Séances Récentes (Contexte)
-
-{recent_workouts if recent_workouts else "_Historique non disponible_"}
-
----
-"""
-        # Ajouter feedback athlète si disponible
-
-        if athlete_feedback:
-            feedback_text = self.format_athlete_feedback(athlete_feedback)
-            if feedback_text:
-                prompt += f"""
-## 💭 Retour Athlète (Ressenti Subjectif)
-
-{feedback_text}
-
-**Important** : Ce retour subjectif enrichit l'analyse objective des métriques. Croiser les deux perspectives pour une analyse complète.
-
----
-"""
-        prompt += """
-
-## Demande d'Analyse
-
-En tant qu'assistant coach, analyse cette séance avec un regard factuel et technique.
-
-**Critères d'analyse :**
-1. Factuel, basé uniquement sur les métriques disponibles
-2. Évaluer qualité via découplage (<7.5% = validé)
-3. Contextualiser avec TSB pré-séance et sommeil
-4. Identifier patterns (Sweet-Spot, Endurance, VO2, etc.)
-5. **Intégrer le feedback athlète** (section "Feedback Athlète") s'il est présent - ressenti général (1-5, 1=Excellent 5=Mauvais) et notes textuelles (avec système de fallback: description activité en priorité, wellness comments si vide)
-6. Recommandations concrètes basées sur les données ET le ressenti
-
-**Gestion des données manquantes (activités Strava) :**
-- Si puissance = 0W : Indiquer "_Données non disponibles (source Strava)_"
-- Si découplage = N/A : Analyser sur base FC/durée/TSS uniquement
-- Mentionner les limites de l'analyse dans "Points d'Attention"
-- Suggérer vérification manuelle sur Intervals.icu web si critique
-
----
-
-**Fournis ton analyse dans ce format EXACT (markdown) :**
-
-### {act['name']}
-ID : {act['id']}
-Date : {act['date']}
-
-#### Métriques Pré-séance
-- CTL : {w_pre['ctl']:.0f}
-- ATL : {w_pre['atl']:.0f}
-- TSB : {w_pre['tsb']:+.0f}
-- Sommeil : {sleep_hours:.1f}h
-
-#### Exécution
-- Durée : {act['duration_min']}min
-- IF : {act['intensity']:.2f}
-- TSS : {self.safe_format_metric(act.get('tss'), '.0f', '')}
-- Puissance moyenne : {self.safe_format_metric(avg_power, '.0f', 'W')}
-- Puissance normalisée : {self.safe_format_metric(normalized_power, '.0f', 'W')}
-- Cadence moyenne : {self.safe_format_metric(avg_cadence, '.0f', 'rpm')}
-- FC moyenne : {self.safe_format_metric(avg_hr, '.0f', 'bpm')}
-- Découplage : {decoupling_str}
-
-#### Exécution Technique
-[2-3 phrases sur validation zone, qualité technique, cohérence métriques]
-
-#### Charge d'Entraînement
-[2 phrases sur TSS, TSB, implications]
-
-#### Validation Objectifs
-- ✅/❌ [Critère 1]
-- ✅/❌ [Critère 2]
-- ✅/❌ [Critère 3 si pertinent]
-
-#### Points d'Attention
-- [Point 1]
-- [Point 2 si pertinent]
-
-#### Recommandations Progression
-1. [Recommandation 1]
-2. [Recommandation 2]
-
-#### Métriques Post-séance
-- CTL : {w_post['ctl']:.0f}
-- ATL : {w_post['atl']:.0f}
-- TSB : {w_post['tsb']:+.0f}
-
----
-
-**IMPORTANT :**
-- Générer UNIQUEMENT le bloc markdown ci-dessus
-- Pas de texte explicatif avant ou après
-- Pas de bloc de code (````markdown)
-- Format directement insérable dans workouts-history.md
-- Être concis et factuel
-
----
-
-Génère maintenant l'entrée d'analyse.
-"""
-        return prompt
-
-    def copy_to_clipboard(self, text):
-        """Copy le texte dans le presse-papier macOS."""
-        try:
-            process = subprocess.Popen(
-                ["pbcopy"], stdin=subprocess.PIPE, stdout=subprocess.PIPE, stderr=subprocess.PIPE
-            )
-            process.communicate(text.encode("utf-8"))
-            return True
-        except Exception as e:
-            print(f"⚠️  Erreur lors de la copie dans le presse-papier : {e}")
-            return False
 
 
 def analyze_batch(api, unanalyzed_activities, generator, state, project_root):
