+++ magma_cycling/workflows/prompt/__init__.py
@@ -1,0 +1,1 @@
+"""PromptGenerator mixin modules for god class decomposition."""
